freedom-mee_header-generator: Handle multiple clocks
The HiFive1 has a more complicated clocking system that the QEMU and
FPGA targets do.  This commit allows the MEE header file generator to
generate enough information to inform the MEE about these nested
clocking systems.

The actual generated output is now very tightly coupled to the MEE
implementation, which isn't ideal.  I'm not sure quite how else to deal
with this, though.

There's an additional test set that targets the HiFive1 device tree that
we're actually going to send out (or at least, close to it).

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>