Script to generate dts.fixup file for Zephyr RTOS
While Zephyr RTOS defines their DTS-based code generation interface for
drivers, they currently require a dts.fixup file to redefine the
generated macros to the symbols consumed by the drivers themselves. This
script generates that dts.fixup file based on the board's DTS.

The current roadmap for Zephyr puts the new DTS code generation as a
milestone for Zephyr 1.14, so until that support is released, the
Freedom SDK will need to generate these fixup files for creating a port
of Zephyr to a new board.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>