Add support for rv32imafdc cores
The default set of multilibs on RISC-V GCC implementations doesn't
include a sane setting for rv32imafdc/ilp32.  While this is arguably a
bug (it could be implemented with the rv32imac/ilp32 library), it's also
probably best to be mapping these to rv32imafdc/ilp32f as users probably
want to take advantage of their F registers when calling functions.

In the long run we should fix the multilib generator to provide a
complete set of multilibs, but for now we'll have to live with these
lookup tables.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>