If there is no explicit external bus then match simple-bus
This is a bit of a hack, but we have some device trees floating around
that don't have proper device tree entries for their test memories.
Matching on simple-bus alone isn't enough, as we really need to only
look for empty simple busses, but we should be able to get away with
this for now.

Signed-off-by: Palmer Dabbelt <palmer@sifive.com>