Fix duplicated index for filter condition
In some functions of CPU and PLIC which generated in the metal.h
header file, the if-else condtions are duplicated such as 'idx == 0'
be checked twice.

In PLIC generator, don't use 'else-if' conditional to filter the
max_interrupts condition. It causes dosen't generate tail curly
brackets when the value of interrupts-extended property specifying
only one contexts.

In CPU generator, don't use 'else-if' conditional to filter the num_cpus
condition. It causes dosen't generate tail curly brackets when there is
only one cpu node in dts.

Signed-off-by: Zong Li <zong.li@sifive.com>