Fix target/source symbols for data, itim, bss
Instead of hoping that the target and source symbols for the beginning
of the data, itim, and bss sections match, explicitly set them to the
LOADADDR and ADDR of the section. This fixes bugs where the linker
decided that, for example, metal_segment_data_source_start doesn't have
to match exactly the load address of the data segment, causing corrupted
data in the running program.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>