Require metal,entry to determine reset vector
The value of the metal,entry property will the the following
prop-encoded-array:

 1. Reference to the memory node
 2. Index into the list of tuples in the `reg` or `ranges` property
 3. Offset into the address range of the tuple

As a result we can get rid of all Arty strategies, because selecting
the testram or SPI as reset vector is now the responsibility of the
Devicetree, not the linker script generator.

The memory only strategy is also obsolete, because it's identical to
the chosen strategy.

All remaining strategies always look for metal,entry when determining
reset vector.

Signed-off-by: Nathaniel Graff <nathaniel.graff@sifive.com>